transaction control

+++ resources/transaction.py
@@ -16,11 +16,21 @@ class Transaction(MethodView):
     @blp.response(200, TransactionSchema)
     def get(self, transaction_id):
         transaction = TransactionModel.query.get_or_404(transaction_id)
+        current_user_id = get_jwt_identity()
+
+        if transaction.group.user_id != current_user_id:
+            abort(403, message="You are not authorized to view or modify this transaction.")
+
         return transaction
 
     @jwt_required()
     def delete(self, transaction_id):
         transaction = TransactionModel.query.get_or_404(transaction_id)
+        current_user_id = get_jwt_identity()
+
+        if transaction.group.user_id != current_user_id:
+            abort(403, message="You are not authorized to delete this transaction.")
+
         db.session.delete(transaction)
         db.session.commit()
         return {"message": "Transaction deleted."}
@@ -30,12 +40,15 @@ class Transaction(MethodView):
     @blp.response(200, TransactionSchema)
     def put(self, transaction_data, transaction_id):
         transaction = TransactionModel.query.get(transaction_id)
+        current_user_id = get_jwt_identity()
 
         if transaction:
+            if transaction.group.user_id != current_user_id:
+                abort(403, message="You are not authorized to update this transaction.")
             transaction.price = transaction_data["price"]
             transaction.description = transaction_data["description"]
         else:
-            transaction = TransactionModel(id=transaction_id, **transaction_data)
+            abort(404, message="Transaction not found.")
 
         db.session.add(transaction)
         db.session.commit()
@@ -48,12 +61,16 @@ class TransactionList(MethodView):
     @jwt_required()
     @blp.response(200, TransactionSchema(many=True))
     def get(self):
-        return TransactionModel.query.all()
+        user_id = get_jwt_identity()
+        return TransactionModel.query.filter_by(user_id=user_id).all()
 
     @jwt_required()
     @blp.arguments(TransactionSchema)
     @blp.response(201, TransactionSchema)
     def post(self, transaction_data):
+        user_id = get_jwt_identity()
+        transaction_data["user_id"] = user_id
+
         transaction = TransactionModel(**transaction_data)
 
         try:
